refactor(chain-leader): move transaction selection into module
